third_party/WebKit/LayoutTests/media/track/opera/interfaces/VTTCue/getCueAsHTMLCrash.html

<!DOCTYPE html>
<html>
<div>Test passes if it does not induce a crash.</div>
<script src=../../../../../resources/testharness.js></script>
<script src=../../../../../resources/testharnessreport.js></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

test(function(){
    var c1 = new VTTCue(0, 1, '<c>x\0');
    window.frag = c1.getCueAsHTML();
}, document.title+', creating the cue');

test(function(){
    assert_false(frag.childNodes[0].hasChildNodes(), 'hasChildNodes()');
}, document.title+', >');

test(function(){}, document.title+', ');
test(function(){}, document.title+', x\\0');
</script>
</html>
